Renaming and moving funcs and files
