+++ src/dbt_bouncer/checks/catalog/check_catalog_sources.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING, List, Literal
 
-from dbt_bouncer.conf_validator_base import BaseCheck
+from dbt_bouncer.check_base import BaseCheck
 
 if TYPE_CHECKING:
     from dbt_bouncer.parsers import DbtBouncerCatalogNode, DbtBouncerSource
