+++ src/dbt_bouncer/checks/manifest/check_exposures.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, List, Literal
 
 from pydantic import Field
 
-from dbt_bouncer.conf_validator_base import BaseCheck
+from dbt_bouncer.check_base import BaseCheck
 
 if TYPE_CHECKING:
     import warnings
