+++ src/dbt_bouncer/checks/manifest/check_lineage.py
@@ -3,7 +3,7 @@
 import re
 from typing import TYPE_CHECKING, List, Literal
 
-from dbt_bouncer.conf_validator_base import BaseCheck
+from dbt_bouncer.check_base import BaseCheck
 
 if TYPE_CHECKING:
     from dbt_bouncer.parsers import DbtBouncerManifest, DbtBouncerModel
