+++ src/dbt_bouncer/checks/manifest/check_macros.py
@@ -14,7 +14,7 @@ with warnings.catch_warnings():
 import jinja2
 from jinja2_simple_tags import StandaloneTag
 
-from dbt_bouncer.conf_validator_base import BaseCheck
+from dbt_bouncer.check_base import BaseCheck
 
 
 class TagExtension(StandaloneTag):
