+++ src/dbt_bouncer/checks/run_results/check_run_results.py
@@ -2,7 +2,7 @@
 
 from typing import TYPE_CHECKING, Literal
 
-from dbt_bouncer.conf_validator_base import BaseCheck
+from dbt_bouncer.check_base import BaseCheck
 
 if TYPE_CHECKING:
     from dbt_bouncer.parsers import DbtBouncerRunResult
