+++ src/dbt_bouncer/utils.py
@@ -11,7 +11,6 @@ from functools import lru_cache
 from pathlib import Path
 from typing import Any, List, Mapping
 
-import toml
 import yaml
 
 
@@ -87,59 +86,6 @@ def flatten(structure: Any, key: str = "", path: str = "", flattened=None):
     return flattened
 
 
-def get_config_file_path(
-    config_file: str,
-    config_file_source: str,
-) -> Path:
-    """Get the path to the config file for dbt-bouncer. This is fetched from (in order):
-
-    1. The file passed via the `--config-file` CLI flag.
-    2. A file named `dbt-bouncer.yml` in the current working directory.
-    3. A `[tool.dbt-bouncer]` section in `pyproject.toml` (in current working directory or parent directories).
-
-    Returns:
-        Path: Config file for dbt-bouncer.
-
-    Raises:
-        RuntimeError: If no config file is found.
-
-    """  # noqa: D400, D415
-    logging.debug(f"{config_file=}")
-    logging.debug(f"{config_file_source=}")
-
-    if config_file_source == "COMMANDLINE":
-        logging.debug(f"Config file passed via command line: {config_file}")
-        return Path(config_file)
-
-    if config_file_source == "DEFAULT":
-        logging.debug(f"Using default value for config file: {config_file}")
-        config_file_path = Path.cwd() / config_file
-        if config_file_path.exists():
-            return config_file_path
-
-    # Read config from pyproject.toml
-    logging.info("Loading config from pyproject.toml, if exists...")
-    if (Path().cwd() / "pyproject.toml").exists():
-        pyproject_toml_dir = Path().cwd()
-    else:
-        pyproject_toml_dir = next(
-            (
-                parent
-                for parent in Path().cwd().parents
-                if (parent / "pyproject.toml").exists()
-            ),
-            None,  # type: ignore[arg-type]
-        )  # i.e. look in parent directories for a pyproject.toml file
-
-        if pyproject_toml_dir is None:
-            logging.debug("No pyproject.toml found.")
-            raise RuntimeError(
-                "No pyproject.toml found. Please ensure you have a pyproject.toml file in the root of your project correctly configured to work with `dbt-bouncer`. Alternatively, you can pass the path to your config file via the `--config-file` flag.",
-            )
-
-    return pyproject_toml_dir / "pyproject.toml"
-
-
 @lru_cache
 def get_check_objects() -> set[str]:
     """Get list of Check* classes and check_* functions.
@@ -161,32 +107,6 @@ def get_check_objects() -> set[str]:
             check_objects.append(f"{check_qual_name}.{obj}")
 
     return set(check_objects)
-
-
-def load_config_file_contents(config_file_path: Path) -> Mapping[str, Any]:
-    """Load the contents of the config file.
-
-    Returns:
-        Mapping[str, Any]: Config for dbt-bouncer.
-
-    Raises:
-        RuntimeError: If the config file type is not supported or does not contain the expected keys.
-
-    """
-    if config_file_path.suffix in [".yml", ".yaml"]:
-        return load_config_from_yaml(config_file_path)
-    elif config_file_path.suffix in [".toml"]:
-        toml_cfg = toml.load(config_file_path)
-        if "dbt-bouncer" in toml_cfg["tool"]:
-            return next(v for k, v in toml_cfg["tool"].items() if k == "dbt-bouncer")
-        else:
-            raise RuntimeError(
-                "Please ensure your pyproject.toml file is correctly configured to work with `dbt-bouncer`. Alternatively, you can pass the path to your config file via the `--config-file` flag.",
-            )
-    else:
-        raise RuntimeError(
-            f"Config file must be either a `pyproject.toml`, `.yaml` or `.yaml file. Got {config_file_path.suffix}."
-        )
 
 
 def load_config_from_yaml(config_file: Path) -> Mapping[str, Any]:
